backend/requirements.txt
# Core Dependencies
google-generativeai>=0.8.0
rich>=13.7.0
python-dotenv>=1.0.0
click>=8.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Production Fi Money MCP Integration
aiohttp>=3.9.0
asyncio-mqtt>=0.16.0

# SAndeep Investment System
google-adk>=0.1.0
google-genai>=0.1.0

# HTTP and API
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Date and Time
python-dateutil>=2.8.0

# JSON and YAML
pyyaml>=6.0.0

# Logging and Monitoring
structlog>=23.1.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Web Socket Support
websockets>=12.0

# Cryptography
cryptography>=41.0.0

# Database (PostgreSQL)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Encryption
pycryptodome>=3.19.0

# Scheduler
apscheduler>=3.10.0

# Authentication and Security
PyJWT>=2.8.0
bcrypt>=4.1.0
email-validator>=2.1.0

# Additional FastAPI dependencies
python-multipart>=0.0.6

# PDF Generation
reportlab>=4.0.0
matplotlib>=3.7.0
jinja2>=3.1.0
weasyprint>=60.0

# PDF Processing and Text Extraction
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-magic>=0.4.27
